requirements.txt
datasets
gradio
ipykernel
matplotlib
numpy
pandas
python-dotenv
requests
seaborn
scikit-learn
torch
tqdm
transformers